共通部分切り出し
