+++ src/function/articles/getQiita.ts
@@ -1,18 +1,6 @@
+import { formatDate } from "@/utils/const";
+import { Article } from "@/utils/types";
 import xml2js from "xml2js";
-
-interface Article {
-  title: string;
-  link: string;
-  pubDate: string;
-}
-
-function formatDate(dateString: string): string {
-  const date = new Date(dateString);
-  const year = date.getFullYear();
-  const month = String(date.getMonth() + 1).padStart(2, "0");
-  const day = String(date.getDate()).padStart(2, "0");
-  return `${year}/${month}/${day}`;
-}
 
 export async function fetchArticles(): Promise<Article[]> {
   try {
